Safari-611.1.10/LayoutTests/fast/text/trak-optimizeLegibility-expected.txt

PASS first.getClientRects().length is 1
PASS third.getClientRects().length is 1
FAIL first.getClientRects()[0].right should be within 4 of 346. Was 364.
PASS successfullyParsed is true

TEST COMPLETE
This test makes sure that you can opt-in to honoring the 'trak' table by specifying text-rendering: optimizeLegibility. The first and third lines below should have the same width. The middle one can (but doesn't have to) have a different width.
hamburgefonstiv

hamburgefonstiv

hamburgefonstiv
